+++ packages/package-manager/src/index.ts
@@ -1,6 +1,7 @@
 export * from './PackageManager';
 export * from './NodePackageManagers';
 export { PnpmPackageManager } from './PnpmPackageManager';
+export { YarnPackageManager } from './YarnPackageManager';
 export * from './CocoaPodsPackageManager';
 export { default as shouldUseYarn } from './utils/shouldUseYarn';
 export { default as isYarnOfflineAsync } from './utils/isYarnOfflineAsync';
